Seperated Login Page and functions From Main Page
Created a Main page that houses front end data without login.
