+++ frontend/package.json
@@ -17,7 +17,7 @@
     "react-bootstrap": "^2.10.9",
     "react-dom": "^19.0.0",
     "react-resizable": "^3.0.5",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^5.0.1",
     "sass": "^1.85.0",
     "web-vitals": "^2.1.4"
   },
@@ -45,5 +45,10 @@
       "last 1 safari version"
     ]
   },
-  "proxy": "http://localhost:4000"
+  "options": {
+    "allowedHosts": [
+      "localhost"
+    ],
+    "proxy": "http://localhost:4000/"
+  }
 }
